Move ArchivePicker folder
- Best for consistency
