+++ assets/src/js/Components/archivePicker/MultiSearchOption.js
@@ -1,6 +1,6 @@
 import {useRef, useState, useEffect, useCallback, useMemo} from '@wordpress/element';
 import classNames from 'classnames';
-import {ACTIONS, useArchivePickerContext} from '../ArchivePicker';
+import {ACTIONS, useArchivePickerContext} from './ArchivePicker';
 
 const {__} = wp.i18n;
 const MAX_SEARCHES = 6;
