+++ assets/src/js/media_archive.js
@@ -1,4 +1,4 @@
-import ArchivePicker from './Components/ArchivePicker';
+import ArchivePicker from './Components/ArchivePicker/ArchivePicker';
 import {createRoot} from '@wordpress/element';
 const rootElement = document.getElementById('archive-picker-root');
